moving error handling out to another file
