+++ src/linker/steps/link.js
@@ -2,7 +2,8 @@ import winston from 'winston'
 import { visit } from '../../visitors/visiting'
 import { filtering } from '../../visitors/commons'
 import { linkeables } from '../definitions'
-import { array, isArray, forAll } from '../../utils/collections'
+import { appendError, createUnresolvedLinkageError, createWrongTypeLinkageError } from '../errors'
+import { isArray, forAll } from '../../utils/collections'
 import { Node } from '../../model'
 
 // winston.level = 'debug'
@@ -23,8 +24,6 @@ const doLink = (node, linkDef) => Object.keys(linkDef).forEach(feature => {
   link(node, feature, linkDef[feature])
 })
 
-// TODO: it should check the linked object with link type restriction.
-/* eslint no-unused-vars:0 */
 const link = (node, feature, linkType) => {
   const refValue = node[feature]
 
@@ -45,50 +44,21 @@ const link = (node, feature, linkType) => {
 
   // check resolved types
   if (node[feature] && !linkType(node[feature])) {
-    appendWrongLinkTypeError(node, feature, linkType)
+    appendError(node, createWrongTypeLinkageError(feature))
   }
 }
 const alreadyLinked = refValue => (
   isArray(refValue) ? forAll(refValue, alreadyLinked) : typeof refValue !== 'string'
 )
 
-const resolveAndLink = (n, feature, value, onResolved) => {
-  const found = findInScope(n, value)
+const resolveAndLink = (node, feature, value, onResolved) => {
+  const found = findInScope(node, value)
   if (found) {
     onResolved(Ref(found))
   } else {
-    appendError(n, feature, value)
+    appendError(node, createUnresolvedLinkageError(feature, value))
   }
 }
-
-export const LINKAGE_ERROR_TYPE = 'LINKAGE'
-
-// TODO: unify code between this two types of errors
-const appendError = (node, feature, ref) => {
-  if (!node.errors) {
-    node.errors = []
-  }
-  node.errors.push({
-    errorType: LINKAGE_ERROR_TYPE,
-    feature,
-    ref,
-    message: `Cannot resolve reference to '${ref}'`
-  })
-}
-
-const appendWrongLinkTypeError = (node, feature, linkType) => {
-  if (!node.errors) {
-    node.errors = []
-  }
-  node.errors.push({
-    errorType: LINKAGE_ERROR_TYPE,
-    feature,
-    // TODO: improve this message
-    message: 'Referencing a wrong type'
-  })
-}
-
-export const isLinkageError = error => error.errorType === LINKAGE_ERROR_TYPE
 
 const findInScope = (node, name) =>
   node && ((node.scope && node.scope[name]) || findInScope(node.parent, name))
